put media download logic into its own file
